iterator dfs, children index 찾기 추가

+++ src/main/java/com/project/todotodo/model/Node.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public abstract class Node {
-    private int nodeId;
+    private long nodeId;
     NodeList nodeList;
 
     private String content;
